Big docs update and little refactor
