+++ docs/reference/nn/functional.rst
@@ -10,7 +10,6 @@ Table of Contents
 
 - :ref:`Loss Functions <nn_func_loss_fns>`
 - :ref:`Regularization Functions <nn_func_reg_fns>`
-- :ref:`Linear Algebra <nn_func_linalg>`
 
 .. _nn_func_loss_fns:
 Loss Functions
@@ -28,16 +27,8 @@ Loss Functions
 Regularization Functions
 ~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. autofunction:: linear_operator_learning.nn.functional.orthn_fro_reg
+.. autofunction:: linear_operator_learning.nn.functional.orthonormal_fro_reg
 
-.. autofunction:: linear_operator_learning.nn.functional.orthn_logfro_reg
-
-.. _nn_func_linalg:
-Linear Algebra Utilities
-~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: linear_operator_learning.nn.linalg.covariance
-
-.. autofunction:: linear_operator_learning.nn.linalg.sqrtmh
+.. autofunction:: linear_operator_learning.nn.functional.orthonormal_logfro_reg
 
 .. footbibliography::
