+++ docs/reference/nn/index.rst
@@ -1,0 +1,14 @@
+.. _nn_index:
+========================
+:code:`lol.nn` namespace
+========================
+
+This namespace loosely follows pytorch's structure. It contains a growing list of functions and modules to learn linear operators using neural-network feature maps.
+
+.. toctree::
+    :hidden:
+
+    nn
+    functional
+    linalg
+    stats
