+++ docs/reference/nn/linalg.rst
@@ -1,0 +1,14 @@
+.. _nn_linalg:
+=====================
+:code:`nn.linalg`
+=====================
+
+.. module:: linear_operator_learning.nn
+
+.. _nn_func_linalg:
+Linear Algebra Utilities
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. autofunction:: linear_operator_learning.nn.linalg.sqrtmh
+
+.. footbibliography::
