+++ docs/reference/nn/nn.rst
@@ -8,9 +8,19 @@
 Table of Contents
 -----------------
 
+- :ref:`Regressors <nn_regressors>`
 - :ref:`Loss Functions <nn_loss_fns>`
 - :ref:`Modules <nn_modules>`
-- :ref:`Regressors <nn_regressors>`
+
+.. _nn_regressors:
+Regressors (see also :ref:`kernel regressors <kernel_regressors>`)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. autofunction:: linear_operator_learning.nn.ridge_least_squares
+
+.. autofunction:: linear_operator_learning.nn.eig
+
+.. autofunction:: linear_operator_learning.nn.evaluate_eigenfunction
 
 .. _nn_loss_fns:
 Loss Functions
@@ -48,14 +58,5 @@ Modules
     :members:
     :exclude-members: __init__, __new__, forward
 
-.. _nn_regressors:
-Regressors
-~~~~~~~~~~
-
-.. autofunction:: linear_operator_learning.nn.ridge_least_squares
-
-.. autofunction:: linear_operator_learning.nn.eig
-
-.. autofunction:: linear_operator_learning.nn.evaluate_eigenfunction
-
 .. footbibliography::
+
