+++ docs/reference/nn/stats.rst
@@ -1,0 +1,20 @@
+.. _nn_stats:
+=====================
+:code:`nn.stats`
+=====================
+
+.. module:: linear_operator_learning.nn
+
+.. _nn_func_stats:
+Statistics
+~~~~~~~~~~
+
+.. autofunction:: linear_operator_learning.nn.stats.covariance
+
+.. autofunction:: linear_operator_learning.nn.stats.cov_norm_squared_unbiased
+
+.. autofunction:: linear_operator_learning.nn.stats.cross_cov_norm_squared_unbiased
+
+.. autofunction:: linear_operator_learning.nn.stats.whitening
+    
+.. footbibliography::
