+++ linear_operator_learning/kernel/__init__.py
@@ -1,4 +1,3 @@
 """Kernel methods entry point."""
 
-from linear_operator_learning.kernel.linalg import *  # noqa
 from linear_operator_learning.kernel.regressors import *  # noqa
