+++ linear_operator_learning/kernel/linalg.py
@@ -1,95 +1,11 @@
 """Linear algebra utilities for the `kernel` algorithms."""
 
-from typing import Literal
 from warnings import warn
 
 import numpy as np
-import scipy.linalg
 from numpy import ndarray
 
-from linear_operator_learning.kernel.structs import EigResult, FitResult
-from linear_operator_learning.kernel.utils import sanitize_complex_conjugates, topk
-
-__all__ = ["eig", "evaluate_eigenfunction"]
-
-
-def eig(
-    fit_result: FitResult,
-    K_X: ndarray,  # Kernel matrix of the input data
-    K_YX: ndarray,  # Kernel matrix between the output data and the input data
-) -> EigResult:
-    """Computes the eigendecomposition of a regressor.
-
-    Args:
-        fit_result (FitResult): Fit result as defined in ``linear_operator_learning.kernel.structs``.
-        K_X (ndarray): Kernel matrix of the input data.
-        K_YX (ndarray): Kernel matrix between the output data and the input data.
-
-
-    Shape:
-        ``K_X``: :math:`(N, N)`, where :math:`N` is the sample size.
-
-        ``K_YX``: :math:`(N, N)`, where :math:`N` is the sample size.
-
-        Output: ``U, V`` of shape :math:`(N, R)`, ``svals`` of shape :math:`R`
-        where :math:`N` is the sample size and  :math:`R` is the rank of the regressor.
-    """
-    # SUV.TZ -> V.T K_YX U (right ev = SUvr, left ev = ZVvl)
-    U = fit_result["U"]
-    V = fit_result["V"]
-    r_dim = (K_X.shape[0]) ** (-1)
-
-    W_YX = np.linalg.multi_dot([V.T, r_dim * K_YX, U])
-    W_X = np.linalg.multi_dot([U.T, r_dim * K_X, U])
-
-    values, vl, vr = scipy.linalg.eig(W_YX, left=True, right=True)  # Left -> V, Right -> U
-    values = sanitize_complex_conjugates(values)
-    r_perm = np.argsort(values)
-    vr = vr[:, r_perm]
-    l_perm = np.argsort(values.conj())
-    vl = vl[:, l_perm]
-    values = values[r_perm]
-
-    rcond = 1000.0 * np.finfo(U.dtype).eps
-    # Normalization in RKHS
-    norm_r = weighted_norm(vr, W_X)
-    norm_r = np.where(norm_r < rcond, np.inf, norm_r)
-    vr = vr / norm_r
-
-    # Bi-orthogonality of left eigenfunctions
-    norm_l = np.diag(np.linalg.multi_dot([vl.T, W_YX, vr]))
-    norm_l = np.where(np.abs(norm_l) < rcond, np.inf, norm_l)
-    vl = vl / norm_l
-    result: EigResult = {"values": values, "left": V @ vl, "right": U @ vr}
-    return result
-
-
-def evaluate_eigenfunction(
-    eig_result: EigResult,
-    which: Literal["left", "right"],
-    K_Xin_X_or_Y: ndarray,
-):
-    """Evaluates left or right eigenfunctions using kernel matrices.
-
-    Args:
-        eig_result: EigResult object containing eigendecomposition results
-        which: String indicating "left" or "right" eigenfunctions
-        K_Xin_X_or_Y: Kernel matrix between initial conditions and input data (for right
-            eigenfunctions) or output data (for left eigenfunctions)
-
-
-    Shape:
-        ``eig_result``: ``U, V`` of shape :math:`(N, R)`, ``svals`` of shape :math:`R`
-        where :math:`N` is the sample size and  :math:`R` is the rank of the regressor.
-
-        ``K_Xin_X_or_Y``: :math:`(N_0, N)`, where :math:`N_0` is the number of inputs to
-        predict and :math:`N` is the sample size.
-
-        Output: :math:`(N_0, R)`
-    """
-    vr_or_vl = eig_result[which]
-    rsqrt_dim = (K_Xin_X_or_Y.shape[1]) ** (-0.5)
-    return np.linalg.multi_dot([rsqrt_dim * K_Xin_X_or_Y, vr_or_vl])
+from linear_operator_learning.kernel.utils import topk
 
 
 def add_diagonal_(M: ndarray, alpha: float):
