+++ linear_operator_learning/nn/__init__.py
@@ -1,5 +1,4 @@
 """Neural network methods entry point."""
 
-from linear_operator_learning.nn.linalg import eig, evaluate_eigenfunction
 from linear_operator_learning.nn.modules import *  # noqa: F403
-from linear_operator_learning.nn.regressors import ridge_least_squares
+from linear_operator_learning.nn.regressors import eig, evaluate_eigenfunction, ridge_least_squares
